requirements.txt
asgiref==3.4.1
attrs==21.4.0
cachetools==4.2.4
certifi==2021.10.8
charset-normalizer==2.0.10
Django==3.2.8
django-auth-ldap==3.0.0
django-cors-headers==3.11.0
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
drf-spectacular==0.20.2
et-xmlfile==1.1.0
google-api-core==2.3.2
google-api-python-client==2.34.0
google-auth==2.3.3
google-auth-httplib2==0.1.0
googleapis-common-protos==1.54.0
httplib2==0.20.2
idna==3.3
inflection==0.5.1
jsonschema==4.3.3
numpy==1.22.0
oauth2client==4.1.3
openpyxl==3.0.9
pandas==1.3.5
protobuf==3.19.2
psycopg2==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyDrive==1.3.1
PyJWT==2.3.0
pyparsing==3.0.6
pyrsistent==0.18.0
python-dateutil==2.8.2
python-ldap==3.4.0
pytz==2021.3
PyYAML==6.0
requests==2.26.0
rsa==4.8
six==1.16.0
sqlparse==0.4.2
uritemplate==4.1.1
urllib3==1.26.8
